Add support for itext8 and .NET
