+++ dotnet/src/dotnetframework/GxPdfReportsCS/PdfReportInterface.cs
@@ -10,7 +10,7 @@ namespace GeneXus.Printer
 		static ILog log = log4net.LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
 
 		string _appPath;
-		IReportHandler _pdfReport;
+		protected IReportHandler _pdfReport;
 
 		public GxReportBuilderPdf(string appPath, Stream outputStream)
 		{
